Added Windows support.
Not entirely certain about the choice to place user configurations
in the "~\Local Settings" directory, but there wasn't much better
place to put it.

And besides, anyone skilled enough to edit it manually can find it
despite it being hidden, and anyone who can't should be using an
automated tool anyway.

It should handle failure situations decently, and works just fine
on my system.

Verified that Posix support has not broken.

No work at all on OSX, mainly because fucking Apple makes it
impossible to virtualize, and I'm not about to go buy a $1000+
computer just to write compatibility code.